fix validator stake part

+++ api/v2/validators/handler.go
@@ -201,14 +201,14 @@ func GetValidatorSlashes(c *gin.Context) {
 }
 
 // Get IDs of active validators
-func getActiveValidatorIDs(explorer *core.Explorer) []uint64 {
+func getActiveValidatorIDs(explorer *core.Explorer) []uint {
 	return explorer.Cache.Get("active_validators", func() interface{} {
 		return explorer.ValidatorRepository.GetActiveValidatorIds()
-	}, CacheBlocksCount).([]uint64)
+	}, CacheBlocksCount).([]uint)
 }
 
 // Get total stake of active validators
-func getTotalStakeByActiveValidators(explorer *core.Explorer, validators []uint64) string {
+func getTotalStakeByActiveValidators(explorer *core.Explorer, validators []uint) string {
 	return explorer.Cache.Get("validators_total_stake", func() interface{} {
 		if len(validators) == 0 {
 			return "0"
